Implement command generation feature with improved error handling and UI updates
