+++ cyber-query-ai-frontend/src/lib/api.ts
@@ -24,11 +24,36 @@ export const generateCommand = async (
     return response.data;
   } catch (error) {
     if (axios.isAxiosError(error)) {
-      throw new Error(
-        error.response?.data?.detail || "Failed to generate command"
-      );
+      // Handle different types of errors more specifically
+      if (error.response) {
+        // Server responded with error status
+        const errorData = error.response.data;
+        if (typeof errorData === "string") {
+          throw new Error(errorData);
+        } else if (errorData?.detail) {
+          throw new Error(
+            typeof errorData.detail === "string"
+              ? errorData.detail
+              : JSON.stringify(errorData.detail)
+          );
+        } else if (errorData?.message) {
+          throw new Error(errorData.message);
+        } else {
+          throw new Error(
+            `Server error: ${error.response.status} ${error.response.statusText}`
+          );
+        }
+      } else if (error.request) {
+        // Request was made but no response received
+        throw new Error(
+          "No response from server. Please check if the backend is running on localhost:8000"
+        );
+      } else {
+        // Something else happened
+        throw new Error(`Request failed: ${error.message}`);
+      }
     }
-    throw new Error("Network error occurred");
+    throw new Error("An unexpected error occurred");
   }
 };
